update: organize module imports
