+++ src/data/types/LocalDbStore.ts
@@ -1,4 +1,4 @@
-import { StoreIndex } from './StoreIndex'
+import { StoreIndex } from '@data/types'
 
 export type LocalDbStore<T> = {
 	name: string
