+++ src/data/types/index.ts
@@ -15,3 +15,7 @@ export * from './Recommended'
 export * from './Screenshots'
 export * from './Platform'
 export * from './LocalDbStore'
+export * from './StoreIndex'
+export * from './Image'
+export * from './DataServiceDictionary'
+export * from './InlineBannerStyle'
