+++ src/services/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './IDataService'
 export * from './IGameService'
+export * from './IApiService'
