+++ src/utils/interfaces/IStylingUtils.ts
@@ -1,5 +1,4 @@
-import { Game }              from '@data/types'
-import { InlineBannerStyle } from '@data/types/InlineBannerStyle.ts'
+import { Game, InlineBannerStyle } from '@data/types'
 
 export interface IStylingUtils {
 	getInlineBanner(game?: Game): InlineBannerStyle
